requirements.txt
# Framework web
flask==3.0.3

# Conexión a PostgreSQL
psycopg2-binary==2.9.9

# Manejo de XML
lxml==5.3.0

# Manejo de variables de entorno
python-dotenv==1.0.1

# Logging avanzado
loguru==0.7.2

# Para llamadas HTTP (ej. integración con DIAN más adelante)
requests==2.32.3

# Opcional para producción (Linux)
gunicorn==21.2.0

# Opcional para pruebas unitarias
pytest==8.2.0

# MongoDB
pymongo==4.6.1

# Generación de PDF
reportlab==4.1.0

